requirements.txt
pydantic-ai-slim==0.0.48
python-dotenv==1.0.0
colorama==0.4.6
